requirements.txt
flask==3.0.0
google-generativeai==0.3.2
torch==2.2.0
Pillow==10.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2
tensorflow==2.15.0
werkzeug==3.0.1